Add materials and fix record-present synchronization

+++ dx12_renderer/src/geometry/VertexBuffer.h
@@ -7,17 +7,17 @@ using namespace Microsoft::WRL;
 class VertexBuffer
 {
 public:
-	VertexBuffer( void* vertices, UINT vertexSize, UINT vertexCount );
+	VertexBuffer( void* vertices, size_t vertexSize, size_t vertexCount );
 	virtual ~VertexBuffer();
 
 	void bind( ComPtr<ID3D12GraphicsCommandList> commandList );
 
-	UINT getVertexCount() const { return m_vertexCount; }
+	size_t getVertexCount() const { return m_vertexCount; }
 
 private:
 	void* m_vertices;
-	UINT m_vertexSize;
-	UINT m_vertexCount;
+	size_t m_vertexSize;
+	size_t m_vertexCount;
 	ComPtr<ID3D12Resource> m_bufferResource;
 	ComPtr<ID3D12Resource> m_intermediateUploadBuffer;
 	D3D12_SUBRESOURCE_DATA m_subResourceData;
